+++ ts/packages/memory/conversation/src/importPodcast.ts
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import * as kpLib from "knowledge-processor";
 import { dateTime, getFileName, readAllText } from "typeagent";
 import {
     Podcast,
@@ -9,48 +8,39 @@ import {
     PodcastMessageMeta,
     assignMessageListeners,
 } from "./podcast.js";
-import { ConversationSettings, IMessage } from "knowpro";
+import { ConversationSettings } from "knowpro";
+import { parseTranscript, timestampMessages } from "./transcript.js";
 
+/**
+ * Parses a podcast transcript consisting of turns in a conversation.
+ * Turn are in the format:
+ *  SPEAKER_NAME: TEXT
+ *  SPEAKER_NAME:
+ *  TEXT
+ * @param transcriptText
+ * @returns A tuple:
+ *  - Array of {@link PodcastMessage}
+ * - {@link Set} of identified participants
+ */
 export function parsePodcastTranscript(
     transcriptText: string,
 ): [PodcastMessage[], Set<string>] {
-    const turnParserRegex = /^(?<speaker>[A-Z0-9 ]+:)?(?<speech>.*)$/;
-    const transcriptLines = getTranscriptLines(transcriptText);
-    const participants = new Set<string>();
-    const messages: PodcastMessage[] = [];
-    let curMsg: PodcastMessage | undefined = undefined;
-
-    for (const line of transcriptLines) {
-        const match = turnParserRegex.exec(line);
-        if (match && match.groups) {
-            let speaker = match.groups["speaker"];
-            let speech = match.groups["speech"];
-            if (curMsg) {
-                if (speaker) {
-                    messages.push(curMsg);
-                    curMsg = undefined;
-                } else if (speech) {
-                    curMsg.addContent("\n" + speech);
-                }
-            }
-            if (!curMsg) {
-                if (speaker) {
-                    speaker = prepareSpeakerName(speaker);
-                    participants.add(speaker);
-                }
-                curMsg = new PodcastMessage(
-                    [speech],
-                    new PodcastMessageMeta(speaker),
-                );
-            }
-        }
-    }
-    if (curMsg) {
-        messages.push(curMsg);
-    }
-    return [messages, participants];
+    return parseTranscript(
+        transcriptText,
+        (speaker, speech) =>
+            new PodcastMessage([speech], new PodcastMessageMeta(speaker)),
+    );
 }
 
+/**
+ *
+ * @param transcriptFilePath Path to a podcast transcript
+ * @param podcastName
+ * @param startDate
+ * @param lengthMinutes
+ * @param settings
+ * @returns
+ */
 export async function importPodcast(
     transcriptFilePath: string,
     podcastName?: string,
@@ -72,59 +62,4 @@ export async function importPodcast(
     const pod = new Podcast(podcastName, messages, [podcastName], settings);
     // TODO: add more tags
     return pod;
-}
-
-/**
- * Text (such as a transcript) can be collected over a time range.
- * This text can be partitioned into blocks. However, timestamps for individual blocks are not available.
- * Assigns individual timestamps to blocks proportional to their lengths.
- * @param turns Transcript turns to assign timestamps to
- * @param startDate starting
- * @param endDate
- */
-export function timestampMessages(
-    messages: IMessage[],
-    startDate: Date,
-    endDate: Date,
-): void {
-    let startTicks = startDate.getTime();
-    const ticksLength = endDate.getTime() - startTicks;
-    if (ticksLength <= 0) {
-        throw new Error(`${startDate} is not < ${endDate}`);
-    }
-    let messageLengths = messages.map((m) => messageLength(m));
-    const textLength: number = messageLengths.reduce(
-        (total: number, l) => total + l,
-        0,
-    );
-    const ticksPerChar = ticksLength / textLength;
-    for (let i = 0; i < messages.length; ++i) {
-        messages[i].timestamp = new Date(startTicks).toISOString();
-        // Now, we will 'elapse' time .. proportional to length of the text
-        // This assumes that each speaker speaks equally fast...
-        startTicks += ticksPerChar * messageLengths[i];
-    }
-
-    function messageLength(message: IMessage): number {
-        return message.textChunks.reduce(
-            (total: number, chunk) => total + chunk.length,
-            0,
-        );
-    }
-}
-
-function getTranscriptLines(transcriptText: string): string[] {
-    return kpLib.split(transcriptText, /\r?\n/, {
-        removeEmpty: true,
-        trim: true,
-    });
-}
-
-function prepareSpeakerName(speaker: string): string {
-    speaker = speaker.trim();
-    if (speaker.endsWith(":")) {
-        speaker = speaker.slice(0, speaker.length - 1);
-    }
-    speaker = speaker.toLocaleLowerCase();
-    return speaker;
 }
